+++ assets/src/scripts/components/graph/state/index.js
@@ -6,7 +6,7 @@ export { getCursor, getCursorDatum, setCursor } from './cursor';
 export * from './layout';
 export * from './options';
 export { getActiveClasses, getChartPoints, getCurrentWaterLevels,
-         getCurrentWaterLevelUnit, getLineSegments } from './points';
+         getCurrentWaterLevelUnit, getExtentX, getLineSegments } from './points';
 export * from './scales';
 
 export default {