+++ assets/src/scripts/components/graph/state/layout.spec.js
@@ -21,26 +21,33 @@ describe('graph component layout state', () => {
     });
 
     it('setViewport and resetViewport works', () => {
-        const viewport = {
-            startDate: new Date('2010-10-10'),
-            endDate: new Date('2012-10-10')
-        };
-        store.dispatch(setViewport(viewport));
-        expect(getViewport(store.getState())).toEqual(viewport);
+        const domain = [
+            new Date('2010-10-10'),
+            new Date('2012-10-10')
+        ];
+        expect(getViewport.resultFunc(null, domain)).toEqual(domain);
+        expect(getViewport.resultFunc(domain, null)).toEqual(domain);
+    });
+
+    it('setViewport and resetViewport works with mock store', () => {
+        store = getMockStore();
+        store.dispatch(setViewport([
+            new Date('2010-10-10'),
+            new Date('2012-10-10')
+        ]));
         store.dispatch(resetViewport());
-        expect(getViewport(store.getState())).toEqual(null);
     });
 
     it('getChartPosition works', () => {
         // Works on container size
         expect(getChartPosition('main')(store.getState())).not.toBe(null);
-        expect(getChartPosition('panner')(store.getState())).not.toBe(null);
+        expect(getChartPosition('brush')(store.getState())).not.toBe(null);
 
         // Works with specific container size
         store.dispatch(setContainerSize({width: 10, height: 20}));
         expect(getContainerSize(store.getState())).toEqual({width: 10, height: 20});
         expect(getChartPosition('main')(store.getState())).not.toBe(null);
-        expect(getChartPosition('panner')(store.getState())).not.toBe(null);
+        expect(getChartPosition('brush')(store.getState())).not.toBe(null);
     });
 
     it('setAxisYBBox and getAxisYBBox works', () => {