+++ CHANGELOG.md
@@ -16,3 +16,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add legend to chart when there are classed points
 - Line gaps drawn on gap of 6 months or more
 - y-axis label set to depth to water, feet below land surface
+- Added zoom behaviors to main chart